pipecat-bot/requirements.txt
# loguru is recommended by Pipecat
# Ref: https://docs.pipecat.ai/deployment/pipecat-cloud/fundamentals/logging#session-logging
loguru ~= 0.7.0

pipecat-ai[openai,silero]
pipecat-ai-bey

python-dotenv ~= 1.0
